# terrasignal/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaymentRow": {
        "properties": {
          "amount_due": {
            "title": "Amount Due",
            "type": "number"
          },
          "amount_paid": {
            "title": "Amount Paid",
            "type": "number"
          },
          "days_late": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Late"
          },
          "due_date": {
            "format": "date",
            "title": "Due Date",
            "type": "string"
          },
          "lease_id": {
            "title": "Lease Id",
            "type": "string"
          },
          "paid_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid Date"
          }
        },
        "required": [
          "due_date",
          "paid_date",
          "amount_due",
          "amount_paid",
          "days_late",
          "lease_id"
        ],
        "title": "PaymentRow",
        "type": "object"
      },
...
      "RiskQueueItem": {
        "properties": {
          "band": {
            "enum": [
              "green",
              "amber",
              "red"
            ],
            "title": "Band",
            "type": "string"
          },
          "baseline_mode": {
            "title": "Baseline Mode",
            "type": "boolean"
          },
          "credit_rating": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credit Rating"
          },
          "industry": {
            "title": "Industry",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "integer"
          },
          "monthly_rent_due": {
            "title": "Monthly Rent Due",
            "type": "number"
          },
          "pd": {
            "title": "Pd",
            "type": "number"
          },
          "prediction_id": {
            "format": "uuid",
            "title": "Prediction Id",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "tenant_name": {
            "title": "Tenant Name",
            "type": "string"
          },
          "top_driver": {
            "title": "Top Driver",
            "type": "string"
          },
          "trend": {
            "description": "Monthly mean days-late, last 12 months",
            "items": {
              "type": "number"
            },
            "title": "Trend",
            "type": "array"
          }
        },
        "required": [
          "prediction_id",
          "tenant_id",
          "tenant_name",
          "industry",
          "credit_rating",
          "pd",
          "band",
          "model_version",
          "baseline_mode",
          "top_driver",
          "monthly_rent_due",
          "trend"
        ],
        "title": "RiskQueueItem",
        "type": "object"
      },
      "RiskScoreRequest": {
        "properties": {
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id"
        ],
        "title": "RiskScoreRequest",
        "type": "object"
      },
      "RiskScoreResponse": {
        "properties": {
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "band": {
            "enum": [
              "green",
              "amber",
              "red"
            ],
            "title": "Band",
            "type": "string"
          },
          "baseline_mode": {
            "title": "Baseline Mode",
            "type": "boolean"
          },
          "drivers": {
            "items": {
              "$ref": "#/components/schemas/ShapDriverOut"
            },
            "title": "Drivers",
            "type": "array"
          },
          "model_version": {
            "title": "Model Version",
            "type": "integer"
          },
          "pd": {
            "description": "Calibrated probability of default within 6 months",
            "title": "Pd",
            "type": "number"
          },
          "prediction_id": {
            "format": "uuid",
            "title": "Prediction Id",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "tenant_name": {
            "title": "Tenant Name",
            "type": "string"
          }
        },
        "required": [
          "prediction_id",
          "tenant_id",
          "tenant_name",
          "pd",
          "band",
          "as_of",
          "model_version",
          "baseline_mode",
          "drivers"
        ],
        "title": "RiskScoreResponse",
        "type": "object"
      },
      "ShapDriverOut": {
        "properties": {
          "feature": {
            "title": "Feature",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "shap": {
            "title": "Shap",
            "type": "number"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "feature",
          "label",
          "value",
          "shap"
        ],
        "title": "ShapDriverOut",
        "type": "object"
      },
      "TenantDetail": {
        "properties": {
          "credit_rating": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credit Rating"
          },
          "industry": {
            "title": "Industry",
            "type": "string"
          },
          "latest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskScoreResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "leases": {
            "items": {
              "$ref": "#/components/schemas/TenantLease"
            },
            "title": "Leases",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "payment_history": {
            "items": {
              "$ref": "#/components/schemas/PaymentRow"
            },
            "title": "Payment History",
            "type": "array"
          },
          "score_history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Score History",
            "type": "array"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "name",
          "industry",
          "credit_rating",
          "latest",
          "payment_history",
          "leases",
          "score_history"
        ],
        "title": "TenantDetail",
        "type": "object"
      },
      "TenantLease": {
        "properties": {
          "asset_class": {
            "title": "Asset Class",
            "type": "string"
          },
          "base_rent_psf": {
            "title": "Base Rent Psf",
            "type": "number"
          },
          "commencement": {
            "format": "date",
            "title": "Commencement",
            "type": "string"
          },
          "expiration": {
            "format": "date",
            "title": "Expiration",
            "type": "string"
          },
          "lease_id": {
            "title": "Lease Id",
            "type": "string"
          },
          "lease_type": {
            "title": "Lease Type",
            "type": "string"
          },
          "property_name": {
            "title": "Property Name",
            "type": "string"
          },
          "submarket": {
            "title": "Submarket",
            "type": "string"
          },
          "term_months": {
            "title": "Term Months",
            "type": "integer"
          },
          "unit_id": {
            "title": "Unit Id",
            "type": "string"
          },
          "unit_rsf": {
            "title": "Unit Rsf",
            "type": "integer"
          }
        },
        "required": [
          "lease_id",
          "unit_id",
          "property_name",
          "submarket",
          "asset_class",
          "commencement",
          "expiration",
          "base_rent_psf",
          "term_months",
          "lease_type",
          "unit_rsf"
        ],
        "title": "TenantLease",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "CRE rent forecasting & tenant default-risk platform.",
    "title": "TerraSignal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/risk/queue": {
      "get": {
        "operationId": "risk_queue_api_v1_risk_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RiskQueueItem"
                  },
                  "title": "Response Risk Queue Api V1 Risk Queue Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Risk Queue",
        "tags": [
          "risk"
        ]
      }
    },
    "/api/v1/risk/score": {
      "post": {
        "operationId": "score_tenant_api_v1_risk_score_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskScoreResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Score Tenant",
        "tags": [
          "risk"
        ]
      }
    },
    "/api/v1/risk/tenants/{tenant_id}": {
      "get": {
        "operationId": "tenant_detail_api_v1_risk_tenants__tenant_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Tenant Detail",
        "tags": [
          "risk"
        ]
      }
    },
...
  }
}